+++ backend/backend_with_jaxrs/src/main/java/com/example/backend_with_jaxrs/services/UserService.java
@@ -23,7 +23,7 @@ public class UserService {
     public User register(User userCredentials) throws CustomException {
         String encodedPassword = bCryptPasswordEncoder.encode(userCredentials.getPassword());
         userCredentials.setPassword(encodedPassword);
-        int id = UserDAO.getInstance().register(userCredentials);
+        Long id = UserDAO.getInstance().register(userCredentials);
 
         return UserDAO.getInstance().getUserById(id);
     }
